+++ CHANGELOG.md
@@ -2,6 +2,10 @@
 
 ## Version 0.9.0-dev
 
+- [Move `ActionHandler` methods to `SequentialActionsPlugin`][90]
+
+[90]: https://github.com/hikikones/bevy-sequential-actions/pull/90
+
 ## Version 0.8.0
 
 - [Update to Bevy 0.11][85]